Begin Work
Begin work on making a cmdhandler that does not require the YML file. The main point of doing so is to make commands into elegant one method afairs. Main todos right now are to go ahead and figure out how to deal with permissions in a dynamic way and move over some commands to this new style as a demo for a pull req